fix: move test-only helpers to brevwick-sdk/testing entry (#3)
Core ESM gzip bundle was 2099 B (51 B over the CLAUDE.md < 2 kB budget)
after the console + screenshot merges from main landed eagerly. Fix at
the root cause by splitting test-only mutators out of the production
bundle:

- New packages/sdk/src/core/registry.ts — pure-data module holding
  DEFAULT_RING_LOADERS, registryState ({ loaders }), and the instances
  Map. No setter functions.
- New packages/sdk/src/testing.ts entry point exports
  __setRingsForTesting / __resetBrevwickRegistry. Mutators live here
  only and never enter the shared chunk index.js imports.
- tsup.config.ts adds src/testing.ts to entry; package.json exports
  adds "./testing".
- Tests switched to import the helpers from the new testing entry.
- Inlined the isBrowser() and instanceKey() single-call helpers and
  tightened the duplicate-createBrevwick warning to claim the final
  bytes needed for the "< 2 kB" threshold.

Measurements (gzip -c packages/sdk/dist/index.js | wc -c):
- Before: 2099 B (over)
- After testing-entry split: 2074 B
- After helper inlining: 2048 B (equal, still over)
- After warning shorten: 2043 B (5 B headroom)

Changeset updated to document the new brevwick-sdk/testing entry
(test-only, not part of the public contract).
