+++ packages/sdk/src/rings/__tests__/network.test.ts
@@ -1,9 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import {
-  __resetBrevwickRegistry,
-  __setRingsForTesting,
-  createBrevwick,
-} from '../../core/client';
+import { createBrevwick } from '../../core/client';
+import { __resetBrevwickRegistry, __setRingsForTesting } from '../../testing';
 import { INTERNAL_KEY, type BrevwickInternal } from '../../core/internal';
 import type { Brevwick, NetworkEntry } from '../../types';
