+++ packages/sdk/tsup.config.ts
@@ -7,7 +7,7 @@ import { defineConfig } from 'tsup';
  * budget mandated by `CLAUDE.md`.
  */
 export default defineConfig({
-  entry: ['src/index.ts'],
+  entry: ['src/index.ts', 'src/testing.ts'],
   format: ['esm', 'cjs'],
   dts: true,
   clean: true,
